internal/server/{api,ui},pkg/client: tighten formatting and structure
This commit cleans up a handful of style nits in the torrent-file
upload changes:

In internal/server/ui, the errMissing* sentinels are hoisted to the
top of the file alongside the type block, the form-encoded fallback
path in Add is dropped (the form's enctype guarantees multipart, so
the fallback handles a scenario that can't happen), the int64 cast
on the multipart memory limit is dropped (ParseMultipartForm already
takes int64), the switch-then-switch dispatch becomes if/else then
switch with a blank line between paragraphs, and the doc comment
above Add is corrected to drop "always".

In internal/server/api, the same redundant int64 cast is dropped and
the missing-file error uses the plain string "missing file part"
instead of "missing %q part" with a literal argument.

In pkg/client, do drops its in-function var() block (forbidden by
the formatting rules) in favour of an early-return for the no-body
case, and the multipart body construction in AddFile is hoisted into
a buildAddFileBody helper so the if-blocks get the blank-line
separation the rules require without crowding the caller.

The matching int64 cast in the client test is also dropped. The
multipartBody helper struct moves down next to requestBody in the
test file's helpers section and gains a type() block wrapper to
match the rest of the package.

+++ pkg/client/client_test.go
@@ -121,7 +121,7 @@ func TestClient_AddFile(t *testing.T) {
 				assert.Equal(t, "multipart/form-data", mediaType)
 				require.NotEmpty(t, params["boundary"])
 
-				require.NoError(t, r.ParseMultipartForm(int64(1<<20)))
+				require.NoError(t, r.ParseMultipartForm(1<<20))
 				file, _, err := r.FormFile("file")
 				require.NoError(t, err)
 				defer file.Close()
